src/main/java/br/com/itaucase/apitranferencia/exception/ManipuladorGlobalExcecoes.java
package br.com.itaucase.apitranferencia.exception;

import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;

@ControllerAdvice
public class ManipuladorGlobalExcecoes  {
    @ExceptionHandler(Exception.class)
    public ResponseEntity<String> tratarTodasExcecoes(Exception ex) {
        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(ex.getMessage());
    }
}